requirements.txt
asgiref==3.9.1
Django==5.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.18
django-cors-headers==4.3.1
django-js-asset==3.1.2
django-mdeditor==0.1.20
djangorestframework==3.15.1
pillow==11.3.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
